Replace repeated test into single test.
